[Optimized math] Move truth table to separate header file
